+++ relstorage/adapters/mysql/stats.py
@@ -20,11 +20,6 @@ from ..stats import AbstractStats
 
 class MySQLStats(AbstractStats):
 
-    def get_object_count(self):
-        """Returns the number of objects in the database"""
-        # do later
-        return 0
-
     def get_db_size(self):
         """Returns the approximate size of the database in bytes"""
         conn, cursor = self.connmanager.open()